Moved graphql_block to graphql_core.
