+++ modules/graphql_block/graphql_block.info.yml
@@ -2,6 +2,7 @@ name: GraphQL Block
 type: module
 description: 'Query Drupal blocks with GraphQL.'
 package: GraphQL
+hidden: true
 core: 8.x
 dependencies:
   - graphql_core
