+++ modules/graphql_core/tests/modules/graphql_block_test/config/install/block.block.stark_content.yml
