+++ modules/graphql_core/tests/modules/graphql_block_test/config/install/block_content.type.basic.yml
