+++ modules/graphql_core/tests/modules/graphql_block_test/config/install/core.entity_form_display.block_content.basic.default.yml
