+++ modules/graphql_core/tests/modules/graphql_block_test/config/install/core.entity_view_display.block_content.basic.graphql.yml
