+++ modules/graphql_core/tests/modules/graphql_block_test/config/install/editor.editor.basic_html.yml
