+++ modules/graphql_core/tests/modules/graphql_block_test/config/install/field.field.block_content.basic.body.yml
