+++ modules/graphql_core/tests/modules/graphql_block_test/config/install/field.storage.block_content.body.yml
