+++ modules/graphql_core/tests/modules/graphql_block_test/config/install/system.theme.yml
